Add PTY backpressure and fix rendering details
PTY capture safety (pty_capture.rs):
- Switch from unbounded mpsc::channel to sync_channel(1024)
- Adds backpressure to prevent OOM from fast-producing subprocesses
- Reader thread blocks when buffer fills, propagating to child process
- Break reader loop on send error for clean shutdown

Paragraph rendering (paragraph.rs):
- Fix local_scroll type: pass as usize instead of casting to u16
- Matches draw_text_span_scrolled signature

StatusLine safety (status_line.rs):
- Replace if/else division with checked_div/checked_rem
- Prevents potential issues with zero spacers
- Format test assertion for clippy

Session documentation:
- Update FIXES_SUMMARY.md with Part 13 fixes
- Update SESSION_TODO.md with completed PTY and link support items

+++ crates/ftui-widgets/src/paragraph.rs
@@ -213,7 +213,7 @@ impl Widget for Paragraph<'_> {
                         span.content.as_ref(),
                         span_style,
                         text_area.right(),
-                        local_scroll as u16,
+                        local_scroll,
                     );
                 } else {
                     draw_text_span(
